contentLength negatif fix

+++ EventLoop.cpp
@@ -6,7 +6,7 @@
 /*   By: oamairi <oamairi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/30 18:32:40 by oamairi           #+#    #+#             */
-/*   Updated: 2026/05/11 15:41:39 by oamairi          ###   ########.fr       */
+/*   Updated: 2026/05/14 13:17:31 by oamairi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -117,9 +117,9 @@ void	EventLoop::run()
 							if (_buffers[_fds[i].fd].find("Content-Length: ") != std::string::npos)
 							{
 								int ligneContentLength = _buffers[_fds[i].fd].find("Content-Length: ");
-								size_t contentLength = std::atoi(_buffers[_fds[i].fd].substr(ligneContentLength + 16).c_str());
+								int contentLength = std::atoi(_buffers[_fds[i].fd].substr(ligneContentLength + 16).c_str());
 								std::string body = _buffers[_fds[i].fd].substr(_buffers[_fds[i].fd].find("\r\n\r\n") + 4);
-								if (contentLength > 0 && body.size() < contentLength)
+								if (contentLength > 0 && (int) body.size() < contentLength)
 									continue;
 							}
 							HttpRequest request;
